enemy impact

+++ app/src/main/java/gradle/model/EpsilonModel.java
@@ -13,7 +13,7 @@ public class EpsilonModel extends Model {
     public static final List<Model> removedItems = new ArrayList<>();
 
     public EpsilonModel() {
-        anchor = new Point2D.Double(100, 100);
+        anchor = new Point2D.Double(300, 200);
         w = Constants.EPSILON_DIAMETER;
         h = Constants.EPSILON_DIAMETER;
         max_speed = Constants.MOVE_SPEED;
